+++ R/tbl_merge.R
@@ -1,18 +1,20 @@
-#' Merge two or more gtsummary regression objects
+#' Merge two or more gtsummary objects
 #'
-#' Merges two or more `tbl_regression`, `tbl_uvregression`, or `tbl_stack`
-#' objects and adds appropriate spanning headers.
+#' Merges two or more `tbl_regression`, `tbl_uvregression`, `tbl_stack`,
+#' or `tbl_summary` objects and adds appropriate spanning headers.
 #'
-#' @param tbls List of gtsummary regression objects
+#' @param tbls List of gtsummary objects to merge
 #' @param tab_spanner Character vector specifying the spanning headers.
-#' Must be same length as `tbls`argument
+#' Must be same length as `tbls` argument
 #' @family tbl_regression tools
 #' @family tbl_uvregression tools
+#' @family tbl_summary tools
 #' @seealso [tbl_stack]
 #' @author Daniel D. Sjoberg
 #' @export
 #' @return A `tbl_merge` object
 #' @examples
+#' # Side-by-side Regression Models
 #' library(survival)
 #' t1 <-
 #'   glm(response ~ trt + grade + age, trial, family = binomial) %>%
@@ -20,73 +22,114 @@
 #' t2 <-
 #'   coxph(Surv(ttdeath, death) ~ trt + grade + age, trial) %>%
 #'   tbl_regression(exponentiate = TRUE)
-#' tbl_merge_ex <-
+#' tbl_merge_ex1 <-
 #'   tbl_merge(
 #'     tbls = list(t1, t2),
 #'     tab_spanner = c("Tumor Response", "Time to Death")
 #'   )
+#' \donttest{
+#' # Descriptive statistics alongside univariate regression, with no spanning header
+#' t3 <-
+#'   trial %>%
+#'   dplyr::select(age, grade, response) %>%
+#'   tbl_summary(missing = "no") %>%
+#'   add_n()
+#' t4 <-
+#'   tbl_uvregression(
+#'     trial %>% dplyr::select(ttdeath, death, age, grade, response),
+#'     method = coxph,
+#'     y = Surv(ttdeath, death),
+#'     exponentiate = TRUE,
+#'     hide_n = TRUE
+#'   )
+#' tbl_merge_ex2 <-
+#'   tbl_merge(tbls = list(t3, t4)) %>%
+#'   as_gt(exclude = "tab_spanner") %>%
+#'   gt::cols_label(stat_0_1 = gt::md("**Summary Statistics**"))
+#' }
 #' @section Example Output:
-#' \if{html}{\figure{tbl_merge_ex.png}{options: width=70\%}}
+#' \if{html}{Example 1}
+#'
+#' \if{html}{\figure{tbl_merge_ex1.png}{options: width=70\%}}
+#'
+#' \if{html}{Example 2}
+#'
+#' \if{html}{\figure{tbl_merge_ex2.png}{options: width=65\%}}
 #'
 tbl_merge <- function(tbls,
-                      tab_spanner = paste0(c("Model "), seq_len(length(tbls)))) {
+                      tab_spanner = paste0(c("Table "), seq_len(length(tbls)))) {
   # input checks ---------------------------------------------------------------
   # class of tbls
   if (!"list" %in% class(tbls)) {
     stop("Expecting 'tbls' to be a list, e.g. 'tbls = list(tbl1, tbl2)'")
   }
 
-  # checking all inputs are class tbl_regression, tbl_uvregression, or tbl_stack
-  if (!map_chr(tbls, class) %in% c("tbl_regression", "tbl_uvregression", "tbl_stack") %>% any()) {
-    stop("All objects in 'tbls' must be class 'tbl_regression', 'tbl_uvregression', or 'tbl_stack'")
+  # checking all inputs are class tbl_regression, tbl_uvregression, tbl_regression, or tbl_stack
+  if (!map_chr(tbls, class) %in%
+      c("tbl_regression", "tbl_uvregression", "tbl_summary", "tbl_stack") %>% all()) {
+    stop(paste("All objects in 'tbls' must be class 'tbl_regression',",
+               "'tbl_uvregression', 'tbl_summary', or 'tbl_stack'"))
   }
 
   # at least two objects must be passed
   tbls_length <- length(tbls)
-  if (tbls_length < 1) stop("Supply 2 or more gtsummary regression objects to 'tbls ='")
+  if (tbls_length < 2) stop("Supply 2 or more gtsummary regression objects to 'tbls ='")
 
   # length of cpannign header matches number of models passed
   if (tbls_length != length(tab_spanner)) {
     stop("'tbls' and 'tab_spanner' must be the same length")
   }
 
+  # for tbl_summary, moving footnote above label column to the first stat_* column
+  tbls <- map_if(
+    tbls,
+    ~ class(.x) == "tbl_summary",
+    function(x){
+      x$table_header$footnote[startsWith(x$table_header$column, "stat_")][1] =
+        x$table_header$footnote[x$table_header$column == "label"][1]
+
+      x$table_header$footnote[x$table_header$column == "label"][1] = list(NULL)
+      return(x)
+    }
+  )
+
   # merging tables -------------------------------------------------------------
   # nesting data by variable (one line per variable), and renaming columns with number suffix
   nested_table <- tbls %>%
     map("table_body") %>%
     imap(function(x, y) {
+      # creating a column that is the variable label
+      group_by(x, .data$variable) %>%
+      mutate(
+        var_label = ifelse(.data$row_type == "label", .data$label, NA)
+      ) %>%
+      tidyr::fill(.data$var_label, .direction = "downup") %>%
+      ungroup() %>%
       rename_at(
-        x,
-        vars(-c("variable", "var_type", "row_type", "label")),
+        vars(-c("variable", "row_type", "var_label", "label")),
         ~ glue("{.}_{y}")
       )
     })
 
-  # THIS IS FROM PR #195, the if-else won't be required when tidyr >=1.0.0 is deps
-  if (tidyr_has_legacy_nest()) {
-    nested_table <- map(
-      nested_table,
-      ~ nest(.x, data = -one_of(c("variable", "var_type")))
-    )
-  } else {
-    nested_table <- map(
-      nested_table,
-      ~ nest(.x, -c("variable", "var_type"))
-    )
-  }
+  # nesting results within variable
+  nested_table <- map(
+    nested_table,
+    ~ nest(.x, data = -one_of(c("variable", "var_label")))
+  )
 
   # merging formatted objects together
   merged_table <-
     nested_table[[1]] %>%
-    rename(
-      table = .data$data
-    )
+    rename(table = .data$data)
 
   # cycling through all tbls, merging results into a column tibble
   for (i in 2:tbls_length) {
     merged_table <-
       merged_table %>%
-      full_join(nested_table[[i]], by = c("variable", "var_type")) %>%
+      full_join(
+        nested_table[[i]],
+        by = c("variable", "var_label")
+      ) %>%
       mutate(
         table = map2(
           .data$table, .data$data,
@@ -101,20 +144,15 @@ tbl_merge <- function(tbls,
           }
         )
       ) %>%
-      select(-c("data"))
+      select(-c("data", "table"), "table")
   }
 
   # unnesting results from within variable column tibbles
-  # THIS IS FROM PR #195, the if-else won't be required when tidyr >=1.0.0 is deps
-  if (tidyr_has_legacy_nest()) {
-    table_body <-
-      unnest(merged_table, "table") %>%
-      select(.data$label, everything())
-  } else {
-    table_body <-
-      unnest(merged_table) %>%
-      select(.data$label, everything())
-  }
+  table_body <-
+    merged_table %>%
+    select(-.data$var_label) %>%
+    unnest("table") %>%
+    select(.data$label, everything())
 
   # stacking all table_header dfs together and renaming
   table_header <-
@@ -122,46 +160,26 @@ tbl_merge <- function(tbls,
       tbls,
       ~ pluck(.x, "table_header") %>%
         mutate(
-          column = paste0(column, "_", .y),
+          column = paste0(.data$column, "_", .y),
           fmt = stringr::str_replace(fmt, stringr::fixed("x$"), paste0("x$tbls[[", .y, "]]$"))
         )
-    )
+    ) %>%
+    # using the identifying columns from first passed object
+    mutate(
+      column = case_when(
+        .data$column == "label_1" ~ "label",
+        .data$column == "variable_1" ~ "variable",
+        .data$column == "row_type_1" ~ "row_type",
+        TRUE ~ .data$column
+      )
+    ) %>%
+    # deleting labels from subsequent merged tables (i >= 2)
+    filter(!startsWith(.data$column, "label_"))
 
   table_header <-
     tibble(column = names(table_body)) %>%
     left_join(table_header, by = "column") %>%
     table_header_fill_missing()
-
-  # creating column footnotes --------------------------------------
-  # creating list of footnote information
-  footnote_abbreviation <- list()
-  footnote_abbreviation[["footnote"]] <-
-    imap(
-      tbl_inputs(tbls),
-      ~ estimate_header(.x$x, .x$exponentiate) %>% attr("footnote")
-    ) %>%
-    unique() %>%
-    compact() %>%
-    unlist() %>%
-    c("CI = Confidence Interval") %>%
-    glue_collapse(sep = ", ")
-
-  footnote_abbreviation[["columns"]] <-
-    imap_lgl(
-      tbl_inputs(tbls),
-      ~ estimate_header(.x$x, .x$exponentiate) %>%
-        attr("footnote") %>%
-        {
-          !is.null(.)
-        }
-    ) %>%
-    which() %>%
-    # if any abbreviations, include estimate_{i} in abbreviation list, otherwise return NULL
-    {
-      switch(length(.) > 0 + 1, paste0("estimate_", .), NULL)
-    } %>%
-    c(paste0("conf.low_", seq_len(tbls_length))) %>%
-    glue_collapse(sep = ", ")
 
   # returning results
   results <- list(
@@ -174,12 +192,6 @@ tbl_merge <- function(tbls,
     call_list = list(tbl_merge = match.call()),
     gt_calls = eval(gt_tbl_merge),
     kable_calls = eval(kable_tbl_merge)
-  )
-
-  # setting column headers
-  results <- modify_header_internal(
-    results,
-    label = "**Characteristic**",
   )
 
   # writing additional gt and kable calls with data from table_header
@@ -207,62 +219,28 @@ gt_tbl_merge <- quote(list(
     glue(),
 
   # Show "---" for reference groups
-  # e.g.fmt_missing(columns = vars(estimate_2, conf.low_2, conf.high_2), rows = (variable == 'trt' & row_type == 'level' & label == \'Drug\'), missing_text = '---') %>%
-  #     fmt_missing(columns = vars(estimate_2, conf.low_2, conf.high_2), rows = (variable == 'grade' & row_type == 'level' & label == \'I\'), missing_text = '---')
   fmt_missing_ref =
     imap(
       tbls,
       function(x, y) {
-        cat_var <-
-          pluck(x, "table_body") %>%
-          filter(
-            .data$var_type == "categorical",
-            .data$row_ref == TRUE,
-            .data$row_type == "level"
-          ) %>%
-          select(c("variable", "row_ref"))
-
-        if (nrow(cat_var) == 0) {
+        # returning NULL for non-regression objects
+        if (!class(x) %in% c("tbl_regression", "tbl_uvregression", "tbl_stack"))
           return(NULL)
-        }
-        map(
-          cat_var$variable,
-          ~ glue(
-            "gt::fmt_missing(",
-            "columns = gt::vars({paste(c('estimate', 'conf.low', 'conf.high'), y, sep = '_', collapse = ', ')}), ",
-            "rows = (variable == '{.x}' & row_type == 'level' & row_ref_{y} == TRUE), ",
-            "missing_text = '---')"
-          )
+        if (class(x) == "tbl_stack" &&
+            !class(x$tbl_regression_list[[1]]) %in% c("tbl_regression", "tbl_uvregression"))
+          return(NULL)
+        # making gt missing code for references
+        glue(
+          "gt::fmt_missing(",
+          "columns = gt::vars({paste(c('estimate', 'ci'), y, sep = '_', collapse = ', ')}), ",
+          "rows = (variable %in% c('{paste(unique(x$table_body$variable), collapse = \"', '\")}') & ",
+          "row_type == 'level' & row_ref_{y} == TRUE), missing_text = '---')"
         )
       }
     ) %>%
-      unlist() %>%
-      compact() %>%
-      glue_collapse_null(),
-
-  # column headers abbreviations footnote
-  footnote_abbreviation = glue(
-    "gt::tab_footnote(",
-    "footnote = '{footnote_abbreviation$footnote}',",
-    "locations = gt::cells_column_labels(",
-    "columns = gt::vars({footnote_abbreviation$columns}))",
-    ")"
-  ),
-
-  # combining conf.low and conf.high to print confidence interval
-  # example result:
-  # cols_merge(col_1 = vars(conf.low_1), col_2 = vars(conf.high_1), pattern = '{1}, {2}') %>%
-  #   cols_merge(col_1 = vars(conf.low_2), col_2 = vars(conf.high_2), pattern = '{1}, {2}')
-  cols_merge_ci =
-    map(
-      seq_len(tbls_length),
-      ~ paste0(
-        "gt::cols_merge(",
-        glue("col_1 = gt::vars(conf.low_{.x}), col_2 = gt::vars(conf.high_{.x}), "),
-        "pattern = '{1}, {2}')"
-      )
-    ) %>%
-      glue_collapse(" %>% "),
+    compact() %>%
+    unlist() %>%
+    glue_collapse_null(),
 
   # indenting levels and missing rows
   tab_style_text_indent = glue(
@@ -279,39 +257,7 @@ gt_tbl_merge <- quote(list(
     glue(
       "gt::tab_spanner(label = '{tab_spanner}', columns = gt::ends_with('_{seq_len(tbls_length)}'))"
     ) %>%
-      glue_collapse(sep = " %>% "),
-
-  # qvalue method footnote
-  footnote_q_method =
-    tbls %>%
-      map(pluck("qvalue_method")) %>%
-      imap_dfr(
-        function(x, y) {
-          if (is.null(x)) {
-            return(tibble(i = y, method = NA_character_, var = paste0("q.value_", i)))
-          }
-          return(tibble(i = y, method = x, var = paste0("q.value_", i)))
-        }
-      ) %>%
-      stats::na.omit() %>%
-      left_join(add_q_method_lookup, by = "method") %>%
-      group_by(method_label) %>%
-      nest() %>%
-      mutate(
-        var_list = map_chr(
-          data,
-          ~ .x$var %>% paste(collapse = ", ")
-        ),
-        gt_call = glue(
-          "gt::tab_footnote(",
-          "footnote = '{method_label}', ",
-          "locations = gt::cells_column_labels(",
-          "columns = gt::vars({var_list}))",
-          ")"
-        )
-      ) %>%
-      pull("gt_call") %>%
-      glue_collapse_null(sep = " %>% ")
+    glue_collapse(sep = " %>% ")
 ))
 
 # kable function calls ------------------------------------------------------------
@@ -323,25 +269,27 @@ kable_tbl_merge <- quote(list(
   #  placeholder, so the formatting calls are performed other calls below
   fmt = NULL,
 
-  # combining conf.low and conf.high to print confidence interval
-  cols_merge_ci =
-    map(
-      seq_len(tbls_length),
-      ~ "dplyr::mutate(conf.low_{.x} = ifelse(is.na(estimate_{.x}), NA, paste0(conf.low_{.x}, \", \", conf.high_{.x})) %>% as.character())" %>%
-        glue()
-    ) %>%
-      glue_collapse_null(sep = " %>% "),
-
   # Show "---" for reference groups
   fmt_missing_ref =
-    map(
-      seq_len(tbls_length),
-      ~ glue(
-        "dplyr::mutate_at(dplyr::vars(estimate_{.x}, conf.low_{.x}), ",
-        "~ dplyr::case_when(row_ref_{.x} == TRUE ~ '---', TRUE ~ .))"
-      )
+    imap(
+      tbls,
+      function(x, y) {
+        # returning NULL for non-regression objects
+        if (!class(x) %in% c("tbl_regression", "tbl_uvregression", "tbl_stack"))
+          return(NULL)
+        if (class(x) == "tbl_stack" &&
+            !class(x$tbl_regression_list[[1]]) %in% c("tbl_regression", "tbl_uvregression"))
+          return(NULL)
+        # making mutate missing code for references
+        glue(
+          "dplyr::mutate_at(dplyr::vars(estimate_{y}, ci_{y}), ",
+          "~ dplyr::case_when(row_ref_{y} == TRUE ~ '---', TRUE ~ .))"
+        )
+      }
     ) %>%
-      glue_collapse_null(sep = " %>% ")
+    compact() %>%
+    unlist() %>%
+    glue_collapse_null()
 ))
 
 
@@ -352,7 +300,7 @@ tbl_inputs <- function(tbl) {
   map(
     tbl,
     function(tbl) {
-      if (class(tbl) %in% c("tbl_regression", "tbl_stack")) {
+      if (class(tbl) %in% c("tbl_regression", "tbl_stack", "tbl_summary")) {
         return(pluck(tbl, "inputs"))
       }
       if (class(tbl) == "tbl_uvregression") {