+++ tests/testthat/test-tbl_merge.R
@@ -11,22 +11,20 @@ t0 <-
     y = response,
     method.args = list(family = binomial),
     exponentiate = TRUE,
-    label = list("trt" ~ "Treatment", "grade" ~ "Grade", "age" ~ "Age"),
   )
 # MVA logistic regression
 t1 <-
   glm(response ~ trt + grade + age, trial, family = binomial) %>%
   tbl_regression(
-    label = list("trt" ~ "Treatment", "grade" ~ "Grade", "age" ~ "Age"),
     exponentiate = TRUE
   )
 # MVA cox regression
 t2 <-
   coxph(Surv(ttdeath, death) ~ trt + grade + age, trial) %>%
   tbl_regression(
-    label = list("trt" ~ "Treatment", "grade" ~ "Grade", "age" ~ "Age"),
     exponentiate = TRUE
   )
+
 
 # tbl_stack adjusted model
 covars <- c("trt", "age")
@@ -41,7 +39,6 @@ coxph(
 ) %>%
   tbl_regression(
     include = .x,
-    label = list("trt" ~ "Treatment", "grade" ~ "Grade", "age" ~ "Age"),
     exponentiate = TRUE
   ))
 
@@ -53,7 +50,6 @@ adj_mods[["grade_mod"]] <- coxph(
   trial
 ) %>%
   tbl_regression(
-    label = list("trt" ~ "Treatment", "grade" ~ "Grade", "age" ~ "Age"),
     exponentiate = TRUE
   )
 
@@ -68,9 +64,25 @@ t4 <-
     tab_spanner = c("UVA Tumor Response", "MVA Tumor Response", "MVA Time to Death", "TTD Adjusted for grade")
   )
 
+t5 <-
+  trial %>%
+  dplyr::select(age, grade, response) %>%
+  tbl_summary(missing = "no") %>%
+  add_n()
+t6 <-
+  tbl_uvregression(
+    trial %>% dplyr::select(ttdeath, death, age, grade, response),
+    method = coxph,
+    y = Surv(ttdeath, death),
+    exponentiate = TRUE,
+    hide_n = TRUE
+  )
+
 test_that("no errors/warnings with standard use", {
   expect_error(t4, NA)
   expect_warning(t4, NA)
+  expect_error(tbl_merge(tbls = list(t5, t6)), NA)
+  expect_warning(tbl_merge(tbls = list(t5, t6)), NA)
 })
 
 test_that("number of rows the same after joining", {
@@ -79,3 +91,14 @@ test_that("number of rows the same after joining", {
   expect_true(nrow(t2$table_body) == nrow(t4$table_body))
   expect_true(nrow(t3$table_body) == nrow(t4$table_body))
 })
+
+test_that("tbl_merge throws errors", {
+  expect_error(tbl_merge(t1), "*")
+  expect_error(tbl_merge(list(mtcars)), "*")
+  expect_error(tbl_merge(tbls = list(t5)), "*")
+  expect_error(tbl_merge(tbls = list(t5, t6), tab_spanner = c("Table")), "*")
+})
+
+
+
+